Refactor CellPositionMap to initialize fields upon construction
